+++ packages/stream_chat_flutter/lib/src/theme/message_input_theme.dart
@@ -75,6 +75,7 @@ class StreamMessageInputThemeData with Diagnosticable {
     this.enableSafeArea,
     this.elevation,
     this.shadow,
+    this.useSystemAttachmentPicker,
   });
 
   /// Duration of the [MessageInput] send button animation
@@ -125,6 +126,18 @@ class StreamMessageInputThemeData with Diagnosticable {
   /// Shadow for the [MessageInput] widget
   final BoxShadow? shadow;
 
+  /// If True, allows you to use the system’s default media picker instead of
+  /// the custom media picker provided by the library. This can be beneficial
+  /// for several reasons:
+  ///
+  /// 1. Consistency: Provides a consistent user experience by using the
+  /// familiar system media picker.
+  /// 2. Permissions: Reduces the need for additional permissions, as the system
+  /// media picker handles permissions internally.
+  /// 3. Simplicity: Simplifies the implementation by leveraging the built-in
+  /// functionality of the system media picker.
+  final bool? useSystemAttachmentPicker;
+
   /// Returns a new [StreamMessageInputThemeData]
   /// replacing some of its properties
   StreamMessageInputThemeData copyWith({
@@ -144,6 +157,7 @@ class StreamMessageInputThemeData with Diagnosticable {
     bool? enableSafeArea,
     double? elevation,
     BoxShadow? shadow,
+    bool? useSystemAttachmentPicker,
   }) {
     return StreamMessageInputThemeData(
       sendAnimationDuration:
@@ -164,6 +178,8 @@ class StreamMessageInputThemeData with Diagnosticable {
       enableSafeArea: enableSafeArea ?? this.enableSafeArea,
       elevation: elevation ?? this.elevation,
       shadow: shadow ?? this.shadow,
+      useSystemAttachmentPicker:
+          useSystemAttachmentPicker ?? this.useSystemAttachmentPicker,
     );
   }
 
@@ -183,6 +199,8 @@ class StreamMessageInputThemeData with Diagnosticable {
       borderRadius: BorderRadius.lerp(a.borderRadius, b.borderRadius, t),
       expandButtonColor:
           Color.lerp(a.expandButtonColor, b.expandButtonColor, t),
+      linkHighlightColor:
+          Color.lerp(a.linkHighlightColor, b.linkHighlightColor, t),
       idleBorderGradient:
           Gradient.lerp(a.idleBorderGradient, b.idleBorderGradient, t),
       inputBackgroundColor:
@@ -196,6 +214,7 @@ class StreamMessageInputThemeData with Diagnosticable {
       enableSafeArea: a.enableSafeArea,
       elevation: lerpDouble(a.elevation, b.elevation, t),
       shadow: BoxShadow.lerp(a.shadow, b.shadow, t),
+      useSystemAttachmentPicker: b.useSystemAttachmentPicker,
     );
   }
 
@@ -221,6 +240,7 @@ class StreamMessageInputThemeData with Diagnosticable {
       enableSafeArea: other.enableSafeArea,
       elevation: other.elevation,
       shadow: other.shadow,
+      useSystemAttachmentPicker: other.useSystemAttachmentPicker,
     );
   }
 
@@ -244,7 +264,8 @@ class StreamMessageInputThemeData with Diagnosticable {
           linkHighlightColor == other.linkHighlightColor &&
           enableSafeArea == other.enableSafeArea &&
           elevation == other.elevation &&
-          shadow == other.shadow;
+          shadow == other.shadow &&
+          useSystemAttachmentPicker == other.useSystemAttachmentPicker;
 
   @override
   int get hashCode =>
@@ -263,7 +284,8 @@ class StreamMessageInputThemeData with Diagnosticable {
       linkHighlightColor.hashCode ^
       elevation.hashCode ^
       shadow.hashCode ^
-      enableSafeArea.hashCode;
+      enableSafeArea.hashCode ^
+      useSystemAttachmentPicker.hashCode;
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
@@ -284,6 +306,8 @@ class StreamMessageInputThemeData with Diagnosticable {
       ..add(ColorProperty('linkHighlightColor', linkHighlightColor))
       ..add(DiagnosticsProperty('elevation', elevation))
       ..add(DiagnosticsProperty('shadow', shadow))
-      ..add(DiagnosticsProperty('enableSafeArea', enableSafeArea));
+      ..add(DiagnosticsProperty('enableSafeArea', enableSafeArea))
+      ..add(DiagnosticsProperty(
+          'useSystemAttachmentPicker', useSystemAttachmentPicker));
   }
 }